invest-ai-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
google-generativeai
email-validator
argon2-cffi
python-multipart
streamlit
requests
pandas
plotly